requirements.txt
Pillow
matplotlib
numpy
opencv-contrib-python
opencv-python-headless
selectivesearch
torch
torchvision
xmltodict
cuda-python
streamlit
